// examples/grpc-web-grpc-gateway/codegen_grpc_web/grpc-web.ts
/**
* This file and any referenced files were automatically generated by @osmonauts/telescope@latest
* DO NOT MODIFY BY HAND. Instead, download the latest proto files for your chain
* and run the transpile command or yarn proto command to regenerate this bundle.
*/

import { grpc } from "@improbable-eng/grpc-web";

export interface UnaryMethodDefinitionishR
  extends grpc.UnaryMethodDefinition<any, any> {
  requestStream: any;
  responseStream: any;
}

export type UnaryMethodDefinitionish = UnaryMethodDefinitionishR;